requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
easyocr==1.7.2
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.1
numpy==2.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
pyclipper==1.3.0.post6
pytesseract==0.3.13
python-bidi==0.6.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
sympy==1.13.1
tifffile==2024.9.20
tokenizers==0.20.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3
